Fix scrutinizer issue.
